split PsProduct out of models
