+++ collectors/product_definitions/tasks.py
@@ -6,7 +6,8 @@ from celery.utils.log import get_task_logger
 from django.utils import timezone
 
 from collectors.framework.models import collector
-from osidb.models import PsContact, PsModule, PsProduct, PsUpdateStream
+from osidb.dmodels import PsProduct
+from osidb.models import PsContact, PsModule, PsUpdateStream
 
 from .core import (
     PRODUCT_DEFINITIONS_URL,
