+++ collectors/product_definitions/tests/test_core.py
@@ -3,8 +3,9 @@ from typing import Type
 import pytest
 from django.db import models
 
+from osidb.dmodels import PsProduct
 from osidb.helpers import get_model_fields
-from osidb.models import PsContact, PsModule, PsProduct, PsUpdateStream
+from osidb.models import PsContact, PsModule, PsUpdateStream
 
 from ..constants import PROPERTIES_MAP
 from ..core import (
