+++ osidb/dmodels/__init__.py
@@ -1,0 +1,3 @@
+from .ps_product import PsProduct
+
+__all__ = ("PsProduct",)
